[jar2xml] Generated XML should be UTF-8.
Would you believe that in Java, a platform that was Unicode since
the beginning, java.io.FileWriter generates output in the "default"
encoding (whatever that may be; conceptually similar to Windows
"ANSI" encoding)? No? Me neither.

Be that as it may, FileWriter generates files in non-UTF-8, which
means that the resulting XML file doesn't validate when run against
the Android API level 10 android.jar file. Oops.

Explicitly generate UTF-8 output.