Add generic bounds information to type/method names.
The goal for jar2xml output is to have an XML representation of the
API exported by a .jar file in a form that allows automated
generation of JNI type and member signatures.

For example, android.widget.AdapterView<T> is declared in Java as:

	public class AdapterView<T extends android.widget.Adapter> extends ViewGroup {
	  // ...
	  public abstract T getAdapter();
	}

However, previously the generated XML declares it as AdapterView<T>
(no `extends`). We need the `extends` so that JNI lookups work
properly, e.g. the JNI signature for AdapterView.getAdapter() is:

	()Landroid/widget/Adapter;

Without knowing the `T extends Adapter` part, we have to assume that
T is java.lang.Object, and thus would instead generate a signature of:

	()Ljava/lang/Object;

This is obviously wrong.

The fix is to embed generic bounds information into the
//class/@name, //interface/@name, and //method/@name attribute value,
so that the bounds may be extracted with full fidelity.