Properly support nested generic parameterized types.
The problem with 1eee9b5a is that it didn't do enough. For example, the
resulting XML would contain:

	java.util.Set<java.util.Map.java.util.Map.Entry<K, V>>

which is laughable.

In order to work around the JRE bug, use a regular expression [0].

[0]
  ``Some people, when confronted with a problem, think
    “I know, I'll use regular expressions.”   Now they have two problems.''
  - @jwz; http://regex.info/blog/2006-09-15/247

In this case I still think it's the most appropriate. If (when?) the JRE bug
is fixed, we won't be fixing the bug anymore (the regex won't match), and the
JRE bug is that it's duplicating the package+type name.