From this version, any failed attempt to get generic type retrieval falls back to non-generic one.
This is to fix issues like bug #8619.

Basically this is proguard bug. It should not expose any obfuscated type names
in its output, not just something that the library code invokes.
Though there are likely too many libraries that will be blocked in this way,
we likely had better live with some workaround. Hence this "fix".