Add deploy and undeploy targets.
Previously, the install and uninstall targets would install things
onto a target device, instead of the traditional semantic of
installing the SDK into a target directory.

We're changing that so that install & uninstall follow the traditional
semantic, so introduce new deploy and undeploy targets for the
previous install/uninstall semantics.